// src/enums/HatchStyle.ts
export enum HatchStyle {
    HS_HORIZONTAL = 0,
    HS_VERTICAL = 1,
    HS_FDIAGONAL = 2,
    HS_BDIAGONAL = 3,
    HS_CROSS = 4,
    HS_DIAGCROSS = 5,
    HS_SOLID = 6,
    HS_DENSE1 = 7,
    HS_DENSE2 = 8,
    HS_DENSE3 = 9,
    HS_DENSE4 = 10,
    HS_DENSE5 = 11,
    HS_DENSE6 = 12,
    HS_GRADIENT = 13
}